+++ cookbooks/base/recipes/main.rb
@@ -27,94 +27,18 @@ package "language-pack-gnome-#{box['lang']}"
 package "aspell-#{box['lang']}"
 package "myspell-#{box['lang']}"
 
-# Virtualization support
-vbox_pkgs = %w[
-  virtualbox-guest-dkms
-  virtualbox-guest-utils
-  virtualbox-guest-x11
-]
-
-case platform
-when "ubuntu"
-  # Virtualization support
-  if virtual_box?
-    vbox_pkgs.each do |pkg|
-      package pkg
-    end
-  end
-
-  # Ubuntu example content
-  package "example-content" do
-    action :purge
-  end
-
-  box['users'].each do |username, usr|
-    execute "#{username}-remove_example_content_file" do
-      command "rm #{usr['home']}/examples.desktop"
-      only_if { ::File.exists? "#{usr['home']}/examples.desktop" }
-    end
-  end
-
-  # NX client for QT
-  package "qtnx"
-
-  # Nautilus plugins
-  package "nautilus-filename-repairer"
-  package "nautilus-open-terminal"
-
-  # Office productivity suite
-  package "lo-menubar" if platform_version == 12.04
-  package "libreoffice-style-galaxy"
-
-  support "libreoffice" do
-    section "base"
-  end
-
-  # Commonly used restricted packages for Ubuntu
-  package "ubuntu-restricted-extras" do
-    notifies :run, "execute[install_additional_extras]", :immediately
-  end
-
-  # Simple foundation for reading DVDs
-  execute "install_additional_extras" do
-    cwd "/usr/share/doc/libdvdread4"
-    command "sh install-css.sh"
-    action :nothing
-  end
-when "mint"
-  # Virtualization support
-  unless virtual_box?
-    vbox_pkgs.each do |pkg|
-      package pkg do
-        action :purge
-      end
-    end
-  end
-
-  # Office productivity suite
-  package "libreoffice-style-galaxy"
-
-  support "libreoffice" do
-    section "base"
-  end
-
-  if platform_version < 15
-    # Commonly used restricted packages for Linux Mint
-    repo_name = "videolan-#{platform_codename}"
-
-    apt_repository repo_name do
-      uri "http://download.videolan.org/pub/debian/stable/ /"
-      distribution ""
-      key "http://download.videolan.org/pub/debian/videolan-apt.asc"
-      action :add
-      not_if { ::File.exists? "#{node['apt']['sources_path']}/#{repo_name}.list" }
-    end
-  end
-end
 
 # Gnome-Do: Quickly perform actions on your desktop
 package "gnome-do"
 package "gnome-do-plugins"
+
+
+# Platform-based tasks
+case platform
+when "mint"   then include_recipe "base::main-mint"
+when "ubuntu" then include_recipe "base::main-ubuntu"
+end
+
 
 # MS Office True Type Fonts
 fonts = data_bag_item('resources', 'fonts')